Refactor report service unit tests
